Add a Lambda function to create a user
This needs better password handling, and we could use KMS to generate random
data for the password [1]. It would be nice to not hard-code the group name
in the Python code as well.

Lambda expects a zipped deployment package, even if there's only a single file
of code. Using archive_file means that Terraform handles that for us.

[1]: http://docs.aws.amazon.com/kms/latest/APIReference/API_GenerateRandom.html