Convert assume-any-role policy to policy document
This makes the policy slightly easier to read, and less prone to JSON syntax
errors. We can't use hyphens in the policy document names because they aren't
valid in the interpolation, so use underscores instead.

`effect` defaults to "Allow", but I've included it in the statement anyway
to avoid doubt.