Add logging to create_user lambda function
Having this in the Python code makes more logging show up in the API response,
and it shows up in CloudWatch logs too.

We need to attach the policy from AWSLambdaBasicExecutionRole to our role as
well, so that Lambda has permissions to create log groups and streams and put
log events. Adding the three permissions from that role to our role in our
policy doesn't seem to work - we don't understand why.