Update password policy to min length 16
We have raised an issue with AWS to follow (or rather, add features allowing us
to follow) the password guidance from NCSC:

  https://www.ncsc.gov.uk/guidance/password-guidance-simplifying-your-approach

AWS doesn't communicate the password policy to users when changing password
which makes the complexity rules particularly unpleasant.

Until then we should follow what we can and remove the complexity requirements
and lower the minimum length.