Remove Unauthorized Activity alarm
Most users of this account only have minimal permissions allowed in the
cross-account-access IAM Group. When these users sign in with the AWS Console
they frequently trigger unauthorized access alarms accidentally, it's very easy
to do so when switching Roles.

For now we'll remove this alarm and only leave the alarm for Root user activity.