Make it possible for a user to view their own page without error
If you view your own IAM User in the console at the moment, you get
errors such as:

    You need permissions
    You do not have the permission required to perform this operation. Ask your administrator to add permissions. Learn more
    User: arn:aws:iam::622626885786:user/philip.potter@digital.cabinet-office.gov.uk is not authorized to perform: iam:GetUser on resource: user philip.potter@digital.cabinet-office.gov.uk

Currently, if a user wants to add or update their MFA credentials,
they have to know that they can ignore this error and just click on
the tab to update security credentials.

This is confusing.  We should have a service which doesn't ask users
to ignore errors.

I worked out that this is the set of permissions which makes the page
work.

In principle we could be more restrictive - eg we could restrict users
to only view group policies for groups they are a member of - but this
is a public repo anyway so I think that's overkill.