Add "GetMFADevice" IAM action to self_manage_iam_user policy
Users are seeing the following error in the IAM console when they have
added a U2F device their gds-users IAM user:

no identity-based policy allows the iam:GetMFADevice action

This is happening because the `self_manage_iam_user` policy that is
attached to each user in gds-users does not permit the GetMFADevice
API call [1], which the AWS console uses to display details about the
FIDO/U2F device. Note, it is not impacting the ability to add or use U2F
devices for authentication, only its display in the console.

Fix this by adding the `iam:GetMFADevice` action to the
`self_manage_iam_user` policy. This will only allow users to get details
about their own U2F devices because it is limited to the user's IAM user
ARN.

This is taken from AWS's own example IAM self-manage policy [2].

1. https://awscli.amazonaws.com/v2/documentation/api/latest/reference/iam/get-mfa-device.html
2. https://docs.aws.amazon.com/IAM/latest/UserGuide/reference_policies_examples_iam_mfa-selfmanage.html